+++ src/CollectionInterface.php
@@ -19,6 +19,6 @@ namespace Rekalogika\Mapper;
  * @extends \IteratorAggregate<TKey,TValue>
  * @extends \ArrayAccess<TKey,TValue>
  */
-interface ArrayInterface extends \ArrayAccess, \IteratorAggregate, \Countable
+interface CollectionInterface extends \ArrayAccess, \IteratorAggregate, \Countable
 {
 }